requirements.txt
astroid			
Click			
colorama		
fastapi			
gunicorn		
h11				
isort			
lazy-object-proxy
mccabe			
pydantic		
pylint			
six				
starlette		
uvicorn			
websockets		
wrapt			
uvloop
langcorn
openai
langchain
pandas
python-dotenv
google-cloud-bigquery
pybigquery
pyarrow
google-cloud-bigquery-storage
networkx
python-jose[cryptography]
passlib[bcrypt]
typing
cloud-sql-python-connector
pymysql
python-multipart
requests
cloud-sql-python-connector[pymysql] 
SQLAlchemy